Renamed application to common application for clarity, added event emitter package
