--- apps/legal-gazette-api/src/modules/application/application.service.interface.ts
@@ -1,9 +1,0 @@
-import { SubmitApplicationDto } from './dto/application.dto'
-
-export interface ILegalGazetteApplicationService {
-  submitApplication(body: SubmitApplicationDto): Promise<void>
-}
-
-export const ILegalGazetteApplicationService = Symbol(
-  'ILegalGazetteApplicationService',
-)
