+++ apps/legal-gazette-api/src/modules/common-application/common-application.service.interface.ts
@@ -1,0 +1,7 @@
+import { SubmitApplicationDto } from './dto/common-application.dto'
+
+export interface ICommonApplicationService {
+  submitApplication(body: SubmitApplicationDto): Promise<void>
+}
+
+export const ICommonApplicationService = Symbol('ICommonApplicationService')
