+++ package.json
@@ -138,6 +138,7 @@
     "@nestjs/cache-manager": "2.0.1",
     "@nestjs/common": "^10.3.10",
     "@nestjs/core": "^10.3.10",
+    "@nestjs/event-emitter": "3.0.1",
     "@nestjs/platform-express": "^10.3.10",
     "@nestjs/sequelize": "^10.0.1",
     "@nestjs/swagger": "^7.1.17",
